[Feat] 복사 버튼 추가 및 파일 분리
